jpa 학습 완료

+++ src/main/java/com/gallery_jwt_jpa/account/AccountMapper.java
@@ -11,6 +11,6 @@ import java.util.List;
 public interface AccountMapper {
    int save(AccountJoinReq req);
    AccountLoginRes findByLoginId(AccountLoginReq req);
-   List<String> findAllRolesByMemberId(int memberId);
+   List<String> findAllRolesByMemberId(long memberId);
 
 }
