+++ src/main/java/com/gallery_jwt_jpa/cart/CartMapper.java
@@ -12,11 +12,11 @@ import java.util.List;
 public interface CartMapper {
     int save(CartPostReq req);
 
-    List<CartGetRes> findAllWithItemByMemberId(int memberId);
+    List<CartGetRes> findAllWithItemByMemberId(long memberId);
 
     int deleteByCartIdAndMemberId(CartDeleteReq req);
 
-    int deleteByMemberId(int memberId);
+    int deleteByMemberId(long memberId);
 
 
 }
