+++ src/main/java/com/gallery_jwt_jpa/order/OrderService.java
@@ -23,7 +23,7 @@ public class OrderService {
 
 //    주문정보 저장
 @Transactional
-    public int saveOrder(OrderPostReq req, int logginedMemberId) {
+    public int saveOrder(OrderPostReq req, long logginedMemberId) {
         // 상품 정보 DB로 부터 가져온다
         List<ItemGetRes> itemList = itemMapper.findAllByIdIn(req.getItemIds());
         log.info("itemList={}", itemList);
@@ -71,7 +71,7 @@ public class OrderService {
         return 1;
     }
 
-    public List<OrderGetRes> findAllByMemberId(int memberId) {
+    public List<OrderGetRes> findAllByMemberId(long memberId) {
         return  orderMapper.findAllByMemberIdOrderByIdDesc(memberId);
     }
