+++ src/main/java/com/gallery_jwt_jpa/order/model/OrderPostReq.java
@@ -12,5 +12,5 @@ public class OrderPostReq {
     private String address;
     private String payment;
     private String cardNumber; //    카드 결제가 아니라면 빈 문자열이 들어옴
-    private List<Integer> itemIds;
+    private List<Long> itemIds;
 }
